fix: remove circular dependencies from tests
