--- packages/scenes/utils/test/index.ts
@@ -1,7 +1,0 @@
-import { toEmitValues } from './toEmitValues';
-import { toEmitValuesWith } from './toEmitValuesWith';
-
-export const matchers = {
-  toEmitValues,
-  toEmitValuesWith,
-};
