+++ packages/scenes/utils/test/updateUrlStateAndSyncState.tsx
@@ -1,6 +1,6 @@
 import { UrlQueryMap } from '@grafana/data';
 import { locationService } from '@grafana/runtime';
-import { UrlSyncManager } from '../../src';
+import { UrlSyncManager } from '../../src/services/UrlSyncManager';
 
 export function updateUrlStateAndSyncState(searchParams: UrlQueryMap, urlManager: UrlSyncManager) {
   locationService.partial(searchParams);
